refactor: share trade interface
