+++ src/components/quote/types.ts
@@ -1,0 +1,8 @@
+export interface FinnhubTrade {
+    p: number; // price
+    s: string; // symbol
+    t: number; // timestamp (ms)
+    v: number; // volume
+    c?: number | null;
+    side?: "bid" | "ask";
+}
